+++ frontend/src/pages/middlewares/restrict-login-signup.tsx
@@ -12,7 +12,7 @@ import { SessionStorageKeys } from "@app/const";
 import { useServerConfig } from "@app/context";
 import { authKeys, fetchAuthToken } from "@app/hooks/api/auth/queries";
 import { setAuthToken } from "@app/hooks/api/reactQuery";
-import { useGtm } from "@app/hooks/useGtm";
+import { GtmHead } from "@app/hooks/useGtm";
 
 const QueryParamsSchema = z.object({
   callback_port: z.coerce.number().optional().catch(undefined),
@@ -22,7 +22,6 @@ const QueryParamsSchema = z.object({
 
 export const AuthConsentWrapper = () => {
   const { config } = useServerConfig();
-  useGtm();
   const [hasConsented, setHasConsented] = useState(() => {
     const consentInfo = sessionStorage.getItem(SessionStorageKeys.AUTH_CONSENT);
     if (!consentInfo) {
@@ -52,6 +51,7 @@ export const AuthConsentWrapper = () => {
 
   return (
     <>
+      <GtmHead />
       {config.authConsentContent && !hasConsented && (
         <div className="bg-opacity-90 fixed inset-0 z-50 flex items-center justify-center bg-mineshaft-700/80">
           <div className="max-h-[80vh] w-4/12 overflow-y-auto rounded-lg bg-bunker-800 p-6 text-white">
